HIIT-030 refactor: security 패키지 리펙토링
- 주석 지우기
- 패키지 이동을 통한 정
