+++ src/main/java/com/hiit/api/security/package-info.java
@@ -1,0 +1,2 @@
+/** Security 설정을 위한 패키지 */
+package com.hiit.api.security;
